Combine battleEventChannel and battleSaga
