+++ packages/shared/src/game/match/combat/index.ts
@@ -1,2 +1,1 @@
-export { battle, startBattle } from "./battleSaga";
-export { BattleEvent, BattleFinishEvent, BattleTurnEvent, BATTLE_FINISH_EVENT, BATTLE_TURN_EVENT } from "./battleEventChannel";
+export { battleSaga, startBattle, BattleEvent, BattleFinishEvent, BattleTurnEvent, BATTLE_FINISH_EVENT, BATTLE_TURN_EVENT } from "./battleEventChannel";
