+++ packages/shared/src/game/match/index.ts
@@ -1,4 +1,3 @@
 export { Match } from "./match";
 
-export { battle, startBattle } from "./combat/battleSaga";
-export { BATTLE_FINISH_EVENT } from "./combat/battleEventChannel";
+export { battleSaga, startBattle, BATTLE_FINISH_EVENT } from "./combat/battleEventChannel";
